hackaton/enfermeria-ursina-v3/requirements.txt
ursina>=5.0
pyyaml>=6.0
Pillow>=10.0
